サーバー起動スクリプトの修正

+++ watch-script.js
@@ -1,36 +1,32 @@
 const fs = require("fs");
-const { exec } = require("child_process");
+const { exec, ChildProcess } = require("child_process");
 
 const buildCommand = "npm run build";
 const startServer = "akashic serve -p 3300 -s nicolive -w -B";
+// akashic-serve-sli v2.17.25 では -w を付けると最初の変更が発生するまでサーバーが起動しない
 
-let lastChangeTime = Date.now();
 let buildProc = /** @type {ChildProcess} */ (null);
-const servProc = run(startServer, true);
+let building = false;
+const serverProc = run(startServer, true);
 
-
-watchChangeFile("./src", () => {
-  const currentTime = Date.now();
-  if (currentTime - lastChangeTime < 1000) return;
-  lastChangeTime = currentTime;
-
-  build();
-});
+watchChangeFile("./src", build);
 
 // watchChangeFile("../akashic-sac/akashic-sac-0.2.3.tgz", () => {
 //   exec("npm run usac")
 //     .on("exit", build);
 // });
 
-
 build();
 
-
 function build() {
+  if (building) return;
+  console.log("ビルド実行");
+
+  building = true;
   buildProc?.kill();
   buildProc = run(buildCommand)
     .on("exit", code => {
-      // if (code === 0) servProc.kill();
+      building = false;
     });
 }
